services/ml/requirements.txt
confluent-kafka
redis
psycopg2-binary
transformers==5.3.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1
